moved permission to mainscreen
